Move `check_operations` to `incrementalmerkletree::testing`
